refactor (Amount) : Amount를 정적 팩토리 메소드로 리팩토링 한다
